+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project follows [semantic versioning](http://semver.org).
 ### UNRELEASED
 * [added] Added `copy` subcommand
 * [added] Added support for xattrs in fuse mount
+* [added] Added support for block/char devices
 * [modified] Also documenting common flags in subcommands
 * [modified] Using repository aliases (**conversion needed**)
 * [modified] Remote path must be absolute